pyproject.toml
[project]
name = "talk2scene"
version = "0.1.0"
description = "Convert audio dialogue into scene events (JSONL) for browser animation"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "redis>=5.0",
    "openai>=1.0",
    "Pillow>=10.0",
    "pydub>=0.25",
    "pyyaml>=6.0",
    "imagehash>=4.3",
    "numpy>=1.24",
]

[project.optional-dependencies]
whisper = [
    "openai-whisper>=20231117",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocs-static-i18n>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
talk2scene = "talk2scene.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["talk2scene"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:ament_flake8 -p no:ament_pep257 -p no:ament_copyright -p no:ament_xmllint -p no:launch_testing -p no:launch_ros -p no:ament_lint"
